backend/requirements.txt
requests
beautifulsoup4
google-genai
python-dotenv
pydantic